requirements.txt
opencv-python
Pillow
numpy
pandas
ipython
streamlit
